+++ fabric/src/main/java/com/jship/bushcraft/fabric/BushcraftFabric.java
@@ -20,19 +20,16 @@ public final class BushcraftFabric implements ModInitializer {
         Bushcraft.init();
 
         ItemStorage.SIDED.registerForBlockEntities(
-                (blockEntity, direction) -> ((SpiritItemStorageImpl) ((SpiritItemStorageProvider) blockEntity)
-                        .getItemStorage(direction)).fabricItemStorage,
-                ModBlockEntities.CHIPPER.get(), ModBlockEntities.CRUCIBLE.get(),
-                ModBlockEntities.DRYING_RACK.get());
+                (blockEntity,
+                        direction) -> ((SpiritItemStorageImpl) ((SpiritItemStorageProvider) blockEntity)
+                                .getItemStorage(direction)).getFabricItemStorage(direction),
+                ModBlockEntities.CHIPPER.get(), ModBlockEntities.CRUCIBLE.get(), ModBlockEntities.DRYING_RACK.get());
         ItemStorage.SIDED.registerForBlockEntities(
-                (blockEntity, direction) -> InventoryStorage.of((WorldlyContainer) blockEntity,
-                        direction),
-                ModBlockEntities.WASHER.get());
+                (blockEntity, direction) -> InventoryStorage.of((WorldlyContainer) blockEntity, direction), ModBlockEntities.WASHER.get());
         FluidStorage.SIDED.registerForBlockEntities(
                 (blockEntity,
                         direction) -> ((SpiritFluidStorageImpl) ((SpiritFluidStorageProvider) blockEntity)
                                 .getFluidStorage(direction)).fabricFluidStorage,
-                ModBlockEntities.CRUCIBLE.get(), ModBlockEntities.TREE_TAP.get(),
-                ModBlockEntities.WASHER.get());
+                ModBlockEntities.CRUCIBLE.get(), ModBlockEntities.TREE_TAP.get(), ModBlockEntities.WASHER.get());
     }
 }