+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 * New GUI screens for Chipper, Crucible, and Washer
-* Chipper and Washer are no AbstractFurnace blocks
+* Chipper and Washer are no longer AbstractFurnace blocks
 * Improved item handling api and fixed some more item handling bugs
+* Added many new recipes